Split up `call` module
